backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
openai
playwright
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
python-dotenv
pgvector
